# Memoize user-defined methods during Her attribute definition