--- a/lib/her/model/attributes.rb
+++ b/lib/her/model/attributes.rb
@@ -186,10 +186,11 @@ module Her
         def attributes(*attributes)
           define_attribute_methods attributes
 
+          user_defined_methods = instance_methods
           attributes.each do |attribute|
             attribute = attribute.to_sym
 
-            unless instance_methods.include?(:"#{attribute}=")
+            unless user_defined_methods.include?(:"#{attribute}=")
               define_method("#{attribute}=") do |value|
                 @attributes[:"#{attribute}"] = nil unless @attributes.include?(:"#{attribute}")
                 self.send(:"#{attribute}_will_change!") if @attributes[:'#{attribute}'] != value
@@ -197,7 +198,7 @@ module Her
               end
             end
 
-            unless instance_methods.include?(:"#{attribute}?")
+            unless user_defined_methods.include?(:"#{attribute}?")
               define_method("#{attribute}?") do
                 @attributes.include?(:"#{attribute}") && @attributes[:"#{attribute}"].present?
               end