@solutus 2015-04-28
this issue slows down to show 100 records up to 10 seconds. Vote to merge it to master