@toddmazierski 2015-05-05
Hi, @marshall-lee!

> Could you provide your benchmark code?

Hmm…not exactly sure to which code you're referring.

  * `Benchmark` is regular old `Benchmark` from from the [standard library](http://ruby-doc.org/stdlib-2.2.2/libdoc/benchmark/rdoc/Benchmark.html)
  * `User.all` in my application corresponds to an endpoint with a large JSON payload (approximately 25 records, and each of them has a nested relation of about 10 records each)

Hope that's useful!

By the way, your solution #337 looks to be a superior one! I'm going to close this pull request.