@marshall-lee 2015-05-05
@toddmazierski I thought maybe you have a complete benchmark code that somehow mocks an API endpoint :smiley: