@marshall-lee 2015-05-05
Because right now I'm trying to achieve this because this issue is not the only one for performance optimization, though it's the most heavy. So having a benchmarking tool for Her would be useful.