@marshall-lee 2015-05-05
Oh, benchmarks here: https://github.com/remiprev/her/pull/337#issuecomment-99233830
@toddmazierski I think you'll be interested.